--- a/crossbeam-channel/Cargo.toml
+++ b/crossbeam-channel/Cargo.toml
@@ -20,10 +20,6 @@ parking_lot = "0.7"
 rand = "0.6"
 smallvec = "0.6.2"
 
-[dependencies.crossbeam-epoch]
-version = "0.7"
-path = "../crossbeam-epoch"
-
 [dependencies.crossbeam-utils]
 version = "0.6.3"
 path = "../crossbeam-utils"