--- a/crossbeam-channel/src/lib.rs
+++ b/crossbeam-channel/src/lib.rs
@@ -347,7 +347,6 @@
 #![warn(missing_docs)]
 #![warn(missing_debug_implementations)]
 
-extern crate crossbeam_epoch;
 extern crate crossbeam_utils;
 extern crate parking_lot;
 extern crate rand;