Use a different queue for unbounded channels.

This gives us:

* Performance improvements (see benchmarks below).
* Lower memory consumption (memory reclamation is not deferred, it's *eager*).
* Fewer dependencies (no more `crossbeam-epoch`).

Before:

```
unbounded_mpmc            Rust crossbeam-channel   0.392 sec
unbounded_mpsc            Rust crossbeam-channel   0.373 sec
unbounded_select_both     Rust crossbeam-channel   0.536 sec
unbounded_select_rx       Rust crossbeam-channel   0.589 sec
unbounded_seq             Rust crossbeam-channel   0.462 sec
unbounded_spsc            Rust crossbeam-channel   0.235 sec
```

After:

```
unbounded_mpmc            Rust crossbeam-channel   0.266 sec
unbounded_mpsc            Rust crossbeam-channel   0.250 sec
unbounded_select_both     Rust crossbeam-channel   0.449 sec
unbounded_select_rx       Rust crossbeam-channel   0.438 sec
unbounded_seq             Rust crossbeam-channel   0.333 sec
unbounded_spsc            Rust crossbeam-channel   0.210 sec
```