@stjepang 2018-12-29
@jeehoonkang Here's a reference implementation: https://github.com/stjepang/queue

This unbounded MPMC queue is very similar to Dmitry Vyukov's [bounded MPMC](http://www.1024cores.net/home/lock-free-algorithms/queues/bounded-mpmc-queue) - it's the same idea, except we have a linked list of blocks rather than one big circular array.

This queue is not 100% lock-free (but neither is `SegQueue`) - there are a few locking sections of code, but they're very small so chances of hitting them are slim. They don't seem to have a big effect on scalability, throughput, nor latency, which is good.

In my benchmarks (those in `crossbeam-channel/benchmarks`), this queue seems to be faster than `MsQueue`, `SegQueue`, and `crossbeam-deque`. I'll benchmark with `crossbeam-circbuf`, too - should be interesting. Eliminating `crossbeam-epoch` is a noticeable perfomance win because pinning incurs one TLS access and one `SeqCst` fence per operation. Tokio benchmarks show better results with this queue than with `SegQueue` and `crossbeam-channel`, too. Finally, latency distribution of operations on this queue is better than with other queues. 

Another benefit of this queue is that memory reclamation is not deferred at all - it's fully eager and there is no concept of garbage! As soon as the last reference to a block is dropped, it gets destroyed. I did some tests and measured lower memory overhead in high concurrency scenarios than with queues using epochs. Hopefully this fixes [the problem](https://github.com/firecracker-microvm/firecracker/pull/677#pullrequestreview-179258729) where Firecracker tests were failing on Amazon's internal CI due to excessive memory use by `crossbeam-channel`.

So how does this avoid epochs? The queue has head and tail indices of type `AtomicUsize`, and there are also head and tail pointers to blocks of type `AtomicPtr<Block<T>>`. Each block has a slot where a message can be stored and each slot has a state with three bits: `READ`, `WRITE`, and `DESTROY`. All bits are initially zero.

To send a message, load `tail.index` and `tail.block`, then do CAS to advance `tail.index` forward. If we succeeded, now we "own" the slot at that index in that block. Now it's safe to dereference the block pointer and write the message into the slot. When done, we do a `fetch_or` to set the `WRITE` bit in the slot's state.

To receive a message, similarly load `head.index` and `head.block`, then do CAS to advance `head.index` forward. If we succeeded, now we "own" the slot at that index in that block. Now it's safe to dereference the block pointer and read the message from the slot (but first wait until `WRITE` is set). When done, we do a `fetch_or` to set the `READ` bit in the slot's state. If we notice that the `DESTROY` bit is set too, we must call `destroy(block)`.

Here comes block destruction. If our receive operation got the last slot in the block, after reading the message and setting the `READ` bit, we must call `destroy(block)`. The `destroy` function iterates over slots in the block and sets the `DESTROY` bit in each slot's state. If we notice during `fetch_or` that the `READ` bit is set too, that means nobody is using the slot anymore. However, if the `READ` bit is not set, that means some other thread is still using the slot, but now that thread became responsible for destroying the block so we just return.

That's the gist of it. I'm omitting a few less important remaining details, but hopefully they make sense when reading the code.

I'm very excited about this queue. Although it doesn't seem super interesting from the CS perspective, in practice it seems to be outperforming pretty much everything else as a general-purpose unbounded MPMC queue.