@jeehoonkang 2018-12-29
Your queue seems very interesting.  I'll look into it real soon.

---

> Although it doesn't seem super interesting from the CS perspective,

AFAICT, everything practical is interesting form the CS perspective :)  To [quote Bertrand Meyer](https://arxiv.org/abs/1712.05078):

> The outcomes that matter in research are not numerous publications, best-paper awards, completed PhD theses, keynote invitations, software tools, citations and other measurable signs of progress. I was after real success, in the sense of changing the way the IT industry develops software.