@stjepang 2018-12-29
I haven't tried with that many cores yet, let's do this!

So I'm thinking about writing a comprehensive benchmark suite for queues similar to `crossbeam-channel/benchmarks`, except better. The benchmarks would be organized as follows.

There are scenarios with non-blocking operations:
*  sequential (send N messages, receive N messages)
* mp (send N messages with T threads)
* mc (receive N messages with T threads)

There are scenarios with blocking operations (we spin+yield when the queue is full or empty):
* spsc (send and receive N messages with 2 threads)
* spmc (send and receive N messages with T+1 threads)
* mpsc (send and receive N messages with T+1 threads)
* mpmc (send and receive N messages with 2T threads)

We measure:
* Total time to complete all operations - that's throughput.
* The time to complete each operation. Then sort the times and print percentiles: 50, 90, 99, 999 - that's latency.

We do this for a bunch of different Ts (number of threads), ranging from 1 to 100 or so.

A few notes about bounded queues:
* In scenarios with non-blocking operations, we set capacity to N so that the queue never gets full.
* In scenarios with blocking operations, we set capacity to 1, 10, 100, and N.

How does this plan sound? What do benchmarks in published papers on concurrent queues do? Should we do something differently?